feat: render footer globally
