+++ src/containers/EditSubjectFrontpage/Subjectpage.tsx
@@ -12,7 +12,6 @@ import { OneColumn } from "@ndla/ui";
 import CreateSubjectpage from "./CreateSubjectpage";
 import EditSubjectpage from "./EditSubjectpage";
 import { usePreviousLocation } from "../../util/routeHelpers";
-import Footer from "../App/components/FooterWrapper";
 import NotFoundPage from "../NotFoundPage/NotFoundPage";
 
 const Subjectpage = () => {
@@ -30,7 +29,6 @@ const Subjectpage = () => {
           <Route path="*" element={<NotFoundPage />} />
         </Routes>
       </OneColumn>
-      <Footer showLocaleSelector={false} />
     </>
   );
 };
