+++ src/containers/StructurePage/StructureContainer.tsx
@@ -35,7 +35,6 @@ import { useUserData } from "../../modules/draft/draftQueries";
 import { useNodes } from "../../modules/nodes/nodeQueries";
 import { createGuard } from "../../util/guards";
 import { getPathsFromUrl, removeLastItemFromUrl } from "../../util/routeHelpers";
-import Footer from "../App/components/FooterWrapper";
 import { useSession } from "../Session/SessionProvider";
 import { useTaxonomyVersion } from "../StructureVersion/TaxonomyVersionProvider";
 import { useLocalStorageBooleanState } from "../WelcomePage/hooks/storedFilterHooks";
@@ -268,7 +267,6 @@ const StructureContainer = ({
         </GridWrapper>
       </PageContent>
       {isTaxonomyAdmin && <VersionSelector />}
-      <Footer showLocaleSelector />
     </ErrorBoundary>
   );
 };
